requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
joblib==1.4.2
numpy==2.1.1
openai==1.48.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pypdfium2==4.30.0
python-dotenv==1.0.1
regex==2024.9.11
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
sniffio==1.3.1
threadpoolctl==3.5.0
tiktoken==0.7.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
